Add Kubernetes and GitHub Action files for EKS deployment
This commit includes new YAML files for setting up a StatefulSet for MariaDB, a Deployment for Wordpress, and a HorizontalPodAutoscaler for Wordpress. It also creates a secret to store the MariaDB root password and a PersistentVolumeClaim for the database. Lastly, a new GitHub Action for EKS deployment is added._eks.yaml file

The deploy_eks.yaml file has been renamed to correct its location under the '.github/workflows' directory. This ensures the EKS deployment workflow is recognized and executed by GitHub actions.