Remove hard-coded secrets and use GitHub secrets instead
The commit removes hard-coded secrets from the configuration files for mariadb and wordpress services. It also deletes the BDD_Secret.yaml file which included these hard-coded secrets. In order to replace the hard-coded secrets, the secrets-sync action is introduced to the deploy workflow to inject secrets from the GitHub secrets store.