Set namespace to 'projet-devops-2023' in Kubernetes YAML files
The namespace 'projet-devops-2023' has been added to all Kubernetes objects in the YAML files. Additionally, the GitHub action 'deploy_eks.yaml' now includes steps to handle the existence of this namespace before deploying the defined objects within it. This move ensures proper resource isolation and access control for the DevOps project in 2023.